requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
openai==1.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
